feat: add delete complete history feature
