+++ src/context/servicesContext/servicesContext.jsx
@@ -17,35 +17,37 @@ const ServiceProvider = ({ children }) => {
   const [state, dispatch] = useReducer(dataReducer, initialDataState);
 
   const getUserCreatedPlaylist = async () => {
-    try {
-      const {
-        data: { playlists }
-      } = await axios.get("/api/user/playlists", {
-        headers: { authorization: authToken }
-      });
-      dispatch({ type: "MANAGE_PLAYLIST", payload: playlists });
-    } catch (error) {
-      console.error("error in getting playlists", error);
+    if (isAuthorized) {
+      try {
+        const {
+          data: { playlists }
+        } = await axios.get("/api/user/playlists", {
+          headers: { authorization: authToken }
+        });
+        dispatch({ type: "MANAGE_PLAYLIST", payload: [...playlists] });
+      } catch (error) {
+        console.error("error in getting playlists", error);
+      }
     }
   };
 
   const getUserHistory = async () => {
-    try {
-      const {
-        data: { history }
-      } = await getHistoryOfUserService(authToken);
-      dispatch({ type: "MANAGE_HISTORY", payload: history });
-    } catch (error) {
-      console.error("error in getting user's history", error);
+    if (isAuthorized) {
+      try {
+        const {
+          data: { history }
+        } = await getHistoryOfUserService(authToken);
+        dispatch({ type: "MANAGE_HISTORY", payload: [...history] });
+      } catch (error) {
+        console.error("error in getting user's history", error);
+      }
     }
   };
 
   useEffect(() => {
-    if (isAuthorized) {
-      getUserCreatedPlaylist();
-      getUserHistory();
-    }
-  });
+    getUserCreatedPlaylist();
+    getUserHistory();
+  }, []);
   return (
     <ServiceContext.Provider value={{ state, dispatch, initialDataState }}>
       {children}
