+++ src/pages/HistoryPage/HistoryPage.jsx
@@ -1,20 +1,42 @@
 import "../HistoryPage/HistoryPage.css";
 import { VideoCard } from "../../components";
 import { useServices } from "../../context/servicesContext/servicesContext";
+import { clearCompleteHistoryOfUserService } from "../../services";
+import { useToast } from "../../custom-hooks/useToast";
+import { useAuth } from "../../context/authContext/authenticationContext";
 
 const HistoryPage = () => {
-  const { state } = useServices();
+  const { state, dispatch } = useServices();
+  const { showToast } = useToast();
+  const { authToken } = useAuth();
+
+  const clearAllHistory = async (e) => {
+    e.preventDefault();
+    try {
+      const {
+        data: { history }
+      } = await clearCompleteHistoryOfUserService(authToken);
+      showToast("History cleared successfully", "success");
+      dispatch({ type: "MANAGE_HISTORY", payload: history });
+    } catch (error) {
+      showToast("Unable to clear history", "error");
+      console.error("Unable to clear all history", error);
+    }
+  };
+
   return (
     <div className="main-content-page">
       <div className="menu-bar">
         <div className="page-title h3 text-bold">My Playlists</div>
         <div>
-          <button className="btn btn-cta">Clear Full History</button>
+          <button className="btn btn-cta" onClick={() => clearAllHistory()}>
+            Clear Full History
+          </button>
         </div>
       </div>
       <div className="history-video-container">
-        {state.history.map(({ _id, title }) => {
-          return <VideoCard key={_id} id={_id} title={title} />;
+        {state.history.map(({ id, title }) => {
+          return <VideoCard key={id} id={id} title={title} />;
         })}
       </div>
     </div>
